requirements.txt
# Airflow providers para integração
apache-airflow-providers-postgres>=5.0.0
apache-airflow-providers-redis>=3.0.0
apache-airflow-providers-apache-spark>=4.0.0
apache-airflow-providers-apache-kafka>=1.0.0
apache-airflow-providers-neo4j>=3.3.0

# Dependências já existentes
pyspark>=4.0.0
minio>=7.1.0
kafka-python-ng>=2.2.2
six>=1.16.0
psycopg2-binary>=2.9.0
redis>=4.5.0
requests>=2.28.0
pandas>=1.5.0
pyarrow>=10.0.0
python-dotenv>=1.1.1
grpcio>=1.74.0
grpcio-status>=1.74.0

# Elasticsearch
elasticsearch==8.11.0

# Apache Pinot
pinotdb>=5.1.0

# StarRocks (MySQL protocol)
pymysql>=1.1.0
